Reorganize geometry/
